requirements.txt
argcomplete==1.8.2
asn1crypto==0.24.0
attrs==19.1.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.5.3
certifi==2019.3.9
chardet==3.0.4
cryptography==2.1.4
cycler==0.10.0
decorator==4.3.2
Django==1.11.11
docx2txt==0.6
EbookLib==0.15
enum34==1.1.6
ephem==3.7.6.0
functools32==3.2.3.post2
GDAL==2.2.2
geoio==1.3.0
idna==2.6
ipaddress==1.0.17
ipython-genutils==0.2.0
jsonschema==3.0.1
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
lxml==4.3.1
matplotlib==2.2.3
nbformat==4.4.0
numpy==1.13.3
opencv-contrib-python==4.1.1.26
Pillow==5.4.1
plotly==3.7.0
pocketsphinx==0.1.3
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-pptx==0.6.5
pytz==2018.3
pyxdg==0.25
requests==2.21.0
retrying==1.3.3
SecretStorage==2.3.1
Shapely==1.6.4.post2
six==1.11.0
SpeechRecognition==3.6.3
sqlparse==0.2.4
subprocess32==3.5.3
textract==1.6.1
tinytools==1.1.0
traitlets==4.3.2
tzwhere==3.0.3
urllib3==1.24.1
virtualenv==15.2.0
xlrd==1.0.0
XlsxWriter==1.1.4
xmltodict==0.12.0
